api/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
psycopg2
flask-bcrypt
flask-session